requirements.txt
dash
pandas
numpy
plotly-express
flask
flask-restful
dash-bootstrap-components